Encoding region generator now includes 4 bit blocks in M1 and M3 codes
Does not account for M3's 4 bit block location changing due to error
correction level. It's currently hard-coded for level L.

It's possible that the idea of blocks can be dropped in the future, in
favour of simply considering the entire encoding region as a single bit
stream. This should be looked into.